Rearchitect clock and timer
These are now seperate entities, and other types no longer respect a
ticking clock, they just get the latest values when it suits them.

This changes the semantics of some tests and models, but most are
unaffected.

The clock is no longer a protocol, it is just a container for delivering
the current time, which can be mocked anyway.