IMPROVE PERFORMANCE FOR REAL
Prototype of how to improve performance of the progress bar rendering.

We should store the current animation state for a given period in  a shared object which can then be read by each progress bar without having to recalculate the state unnecessarily.

This should be implemented in a more testable way, but following the
same rough structure to improve performance.