Support arbitrary-length codes
The length of codes can now be any number, up to UInt16.max in length
(which is crazily long).

Implemented under the hood using BigInt, this seemlessly integrates into
the UI layer, though I'm not sure how groovy it will look in the UI,
it's good we have support for this though.