Resolve detail view invalidation error
The underlying error is actually that we were incorrectly using
@Bindable, when we should have been using @State. The detail view was
getting recreated each time the binding changed, but because we were
initing the view model in the view's init, it was causing the view model
to be recreated incorrectly!

For some reason, this only caused problems on a real device, not the
simulator.