Dynamic keygen algorithm + visibility (#142)
- All key derivation algorithms now have a user-visible description, in
addition to their unique identifier. This allows the user to have
insight into the algorithm to know roughly what it's doing in a
human-readable way.
- Choose the keygen algorithm based on the current optimisation mode,
don't let the user pick this as it's confusing. Most will just the
secure algorithm in production (`RELEASE` builds), but for `DEBUG` we
use the fast algorithm so it doesn't take too long.
- Store the keygen algorithm in the encrypted vault payload, so the key
can be regenerated