ImageDataRenderer renders direct data to UIImage (#146)
- Create new module `ImageTools` that houses image functionality. This
will allow us to use renders directly from `VaultFeed` if needed, rather
than needing to import `CryptoDocumentExporter`, which is kind of
unrelated.
- ImageDataRenderer renders direct data to UIImage. This removes the
intermediate step for encoding to an image format before turning the
data into a UIImage. The advantage of this is that there's a lot less
ambiguity that some given `Data` is just arbitrary data or image data,
like for a `PNG`.
- Update `CryptoDocumentExporter` with this new assumption in mind,
which now imports the renderer from `ImageTools` and encodes shards to
`Data`, removing the PNG encoding step.