Remove InMemoryVaultStore (#179)
- The problem with the `InMemoryVaultStore` is that it tried to
replicate behaviours of the persistent store, but did so improperly. If
we need a production-grade store that is actually stored in memory, the
persistent store should be used with an "in memory only" configuration.
This ensures we get the same querying behaviour etc., but we don't need
to maintain a separate implementation.
- Promote `VaultStoreStub` as the store that should be used if we just
need a test double or simple structure to mock some codes out.