Compress backup payload with LZMA (#185)
- Use LZMA to compress the backup payload to reduce the size of overall
backups. This is really important as we're going to be placing the
backup as QR codes onto paper, so we want the backups to be as small as
possible. LZMA is fine here as backups aren't going to be particularly
large (mainly text), and the encryption process takes a while anyway.
- Uses Apple's default compression helper for LZMA (extension on
NSData), which encodes to the FORMAT_XZ format. This is a standard
format for LZMA.
- CyberChef doesn't seem to give consistent results (likely due to not
supporting the FORMAT_XZ format), so we remove manual verification using
CyberChef, and just used Python instead to verify the manual test cases.
- Resolves #183