Create start of VaultDataModel
This new model will consolidate the storage and observation layers,
providing an abstraction from the store, aimed at the SwiftUI layer.

Still need to add some more tests for tag items, full CRUD support, then
it can start replacing some of this existing logic that's currently
located in some of the viewmodels (e.g. FeedViewModel).

Once this type is ready, the view models can replace their existing
storage interfacing logic with a simple reference to this type. They
will then be sure that all the data they are accessing is up-to-date.