Create `VaultDataModel` UI layer (#220)
- Create a single UI entrypoint to the data in the stores
- This is a single repository that should be injected into necessary
views that provides a consistent view of the available data.
- Vault items and vault tags are housed together in this model as a
change in one of them (e.g. deleting a tag) may need to update the other
as well (e.g. reload items once a given tag is deleted)
- This fixes a few issues where data was inconsistent due to each view
loading it's own copy of the data from the stores. We now just need to
read it from this model to ensure that it's up to date, and there's no
need for views to explicitly fetch their own data.