Rename code 'renderer's to 'codePublisher'
We want a clear distinction between rendering (turning an abstract
OTPAuthCode into a String of digits) and the pipeline that "publishes"
these codes in an efficient way for display in the feed.

We want to create a seperate, simpler way of just rendering an
OTPAuthCode into a String of digits so we can call in on demand, without
having to mess with the existing pipeline of publishers.