Remove BackupKey import and export functionality (#279)
- This is confusing and mostly unnecessary.
- Even if we want to have a vault on 2 devices, sharing the backup key
provides little utility as keygen does not take that long. It could also
confuse users because they might then assume there is syncing enabled
between the two devices, when there is not.
- Users can set the same password on both devices and backups created
from either device will share this password when decrypting. The keygen
salt will be different, but in practice this has almost no effect.