Use custom plugin for linting/formatting (#295)
- Move Makefile to `Vault` package.
- Create a custom plugin for linting/formatting that's a lot simpler,
but still has a lot of control.
- Removes a lot of vendored code that wasn't Swift 6 compatible, so now
we're 100% on Swift 6 (in our own source code).
- Integrate build-time plugin for linting directly in the editor.
- Seperate lint/formatting binary for externally performing external
linting/formatting from the command line.
- If we put linting/formatting *only* as a build tool and run from
there, it requires all sources to be compiled in order to perform this
as it's in the build path. This can slow down things a lot, especially
as a separate binary needs to be created for linting/formatting each
target. This is why having separate binaries for linting/formatting is
useful. SwiftLint is both an external binary (for fast external linting)
and also in the build path, so we get errors as we're editing/compiling.