Task+Race & Task.withTimeout directly return values (#323)
Don't create a wrapper task around Race and Timeout. This makes them
more difficult to work with in practice. Creating a wrapper task when
needed is more appropriate.

This means that races and timeout tasks now run within the task that
originally created them, so respond to cancellation with respect to
those tasks.