Fix TOTP timer updating inconsistently in background (#345)
- This hopefully resolves an issue where, when backgrounded for a while
and returning, a glitchy batch of codes all appear at once within the
preview for TOTP codes.

Actions taken:
- Make the whole of `OTPCodeTimerUpdater` run on the `MainActor` so it
is more provably thread-safe, avoiding the use of `@preconcurrency`.
- When entering the background or becoming inactive, the timer stops
updating codes. It only resumes again on the application becoming
active. This ensures a buffer of codes are not built up in the
background and the codes are only updated when the codes are visible.
- Fix issue where timer bar disappeared when toggling the tags filter
due to the animation preempting the main timer bar animation. Fix this
by adding an ID based on the current search query, forcing the results
to rerender every time the query change.