Import backup that is manually scanned (#369)
- Resolves #310
- This completes the importing story for Vault (for now!)
- Manually scanned codes now create an EncryptedVault which is handed
off directly to the import flow view model, at which point the same flow
is used for manually scanned codes and automatic imports. This includes
handling encryption key collection, decryption and importing and
merging/overriding.
- Prevent swipe to dismiss on the manual scanning view if there is
partially collected state from scans. This is essential because scanning
codes is a lot of work, and we don't want the UI to get dismissed away
mid-scan.