Resolve autofill extension view caching (#391)
- Resolves #386 
- Classic cache invalidation issue.
- Clear view generator cache before the extension displays any content,
to ensure that the previews it is rendering is based off the data most
recently fetched from the store. This was getting out of sync with the
app because these caches were used to render old views, as the extension
had no idea what had changed.
- (This PR includes some attempts and reverts at using a
DarwinNotificationCenter to send notifications between the app and the
extension, which turned out not to be needed.)
- Rename the backing store file, so it has an sqlite extension for easy
introspection. I was using this during debugging to see the changes to
the underlying data store. This is client-breaking, but that's fine as
we're in the unstable 1.0 version.