Fix note preview line, encrypted item searchability (#420)
- When we introduced encrypted items, there was a regression in how the
preview line of the note was displayed. This is now, once again, the
content on the second line of the note. This is populated in the "user
description" field in the note's metadata.
- If the item is encrpyted, we explicitly set this to an empty string,
because the user description is always unencrypted, so we don't want to
leak this on encrypted notes.
- In order to make encrypted items searchable (just the title), make
sure the query engine matches encrypted note titles.